+++ DirectXGameEngine/PostChainManager.cpp
@@ -1,7 +1,7 @@
 #include "PostChainManager.h"
 
 
-PostChainManager::PostChainManager(PostProcess* pp, float width, float height, ID3D11Device* device, ID3D11DeviceContext* deviceContext,ID3D11SamplerState* sampler)
+PostChainManager::PostChainManager(PostProcess* pp, UINT width, UINT height, ID3D11Device* device, ID3D11DeviceContext* deviceContext,ID3D11SamplerState* sampler)
 {
 	postProcess = pp;
 
@@ -13,7 +13,7 @@ PostChainManager::~PostChainManager()
 	delete postProcess;
 }
 
-void PostChainManager::setChain(float width, float height, ID3D11Device* device, ID3D11DeviceContext* deviceContext, ID3D11SamplerState* sampler) {
+void PostChainManager::setChain(UINT width, UINT height, ID3D11Device* device, ID3D11DeviceContext* deviceContext, ID3D11SamplerState* sampler) {
 	Bloom* bloomEffect = new Bloom(width, height, device, deviceContext, sampler);
 	bloomEffect->setBlurAmount(2);
